Improve typeface
OMG please this font is making my eyes bleed. Let's use a more Mozilla-y font stack. 

I ASSUME THIS ALSO NEEDS TO BE HOOKED UP TO GOOGLE FONTS IN THE HEAD OF THE HTML DOC. 

THIS COMMIT DOES NOT DO THAT. Although this commit *does* make the font look better already. Dan, will you look into if the HTML doc needs to be changed??